requirements.txt
openai
streamlit
ffmpeg-python
python-dotenv
yt-dlp
openai-whisper
torch
torchaudio